+++ IoTManager.IDao/IThresholdDao.cs
@@ -6,7 +6,7 @@ namespace IoTManager.IDao
 {
     public interface IThresholdDao
     {
-        Dictionary<String, Tuple<String, int>> GetByDeviceId(String deviceId);
+        Dictionary<String, Tuple<String, double>> GetByDeviceId(String deviceId);
         String Create(ThresholdModel thresholdModel);
         List<ThresholdModel> Get();
     }
