+++ IoTManager.Model/ThresholdModel.cs
@@ -8,7 +8,7 @@ namespace IoTManager.Model
         public String IndexId { get; set; }
         public String DeviceId { get; set; }
         public String Operator { get; set; }
-        public int ThresholdValue { get; set; }
+        public double ThresholdValue { get; set; }
         public DateTime CreateTime { get; set; }
         public DateTime UpdateTime { get; set; }
         public String RuleName { get; set; }
